v.0.2
Implementação: das Funções de Transferência; calculo do erro de rede; calculo do erro de saída.

+++ mlp/src/main/java/com/artificialintelligence/mlp/model/OneHotEncoding.java
@@ -50,19 +50,19 @@ public class OneHotEncoding {
         return classes;
     }
 
-    public float[][] tratarDados(String matrizDadosFront[][], int numSaidas, String[] classes) {
+    public double[][] tratarDados(String matrizDadosFront[][], int numSaidas, String[] classes) {
         int numLinhas = matrizDadosFront.length;
         int numColunas = matrizDadosFront[0].length;
         int posClasse = numColunas-1;
         int qtdeNovasColunas = (numColunas + numSaidas) - 1;
-        float [][]oneHotEnconding = new float[numLinhas-1][qtdeNovasColunas];
+        double [][]oneHotEnconding = new double[numLinhas-1][qtdeNovasColunas];
 
         // - Preencer as classes na Matriz One Hot Encode
         String tempClasse;
         int posAchouClasse;
         for(int i=1, linha=0; i < numLinhas; i++) {
             for (int j=0; j < posClasse; j++) {
-                oneHotEnconding[linha][j] = Float.parseFloat(matrizDadosFront[i][j]);
+                oneHotEnconding[linha][j] = Double.parseDouble(matrizDadosFront[i][j]);
             }
             tempClasse = matrizDadosFront[i][posClasse];
             posAchouClasse = buscaPos(classes, tempClasse);
@@ -77,7 +77,7 @@ public class OneHotEncoding {
     }
 
 
-    public void Exibicao(float[][] dadosOneHotEncoding) {
+    public void Exibicao(double[][] dadosOneHotEncoding) {
         for (int i=0; i<dadosOneHotEncoding.length; i++) {
             for (int j=0; j<dadosOneHotEncoding[0].length; j++) {
                 System.out.printf("%.2f   ", dadosOneHotEncoding[i][j]);
